All mods have float parameters now

+++ Assets/GameLogic/Spells/Modificators/GreatModificator.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using UnityEngine;
 
 public class GreatModificator:  AbstractSpellModificator {
-	public double scaleFactor = 1.0;
+	public float scaleFactor = 1.0f;
 	public override string Name {
 		get {
 			return "great";
